[chore] #18 코드리뷰 반영
